+++ screenpipe-audio/src/bin/screenpipe-audio-forever.rs
@@ -130,6 +130,7 @@ async fn main() -> Result<()> {
                         }
                     }
                 })
+                .await
             }
         })
         .collect();
