+++ screenpipe-audio/src/deepgram/mod.rs
@@ -1,0 +1,3 @@
+mod process_chunk;
+
+pub use process_chunk::transcribe_with_deepgram;
