+++ screenpipe-audio/src/stt.rs
@@ -1,29 +1,19 @@
-use crate::audio_processing::{average_noise_spectrum, spectral_subtraction, write_audio_to_file};
+use crate::audio_processing::write_audio_to_file;
+use crate::deepgram::transcribe_with_deepgram;
 use crate::pyannote::models::{get_or_download_model, PyannoteModel};
 use crate::resample;
 use crate::{
-    audio_processing::normalize_v2,
-    pyannote::{
-        embedding::EmbeddingExtractor,
-        identify::EmbeddingManager,
-        segment::{get_segments, SpeechSegment},
-    },
+    pyannote::{embedding::EmbeddingExtractor, identify::EmbeddingManager},
     vad_engine::{SileroVad, VadEngine, VadEngineEnum, VadSensitivity, WebRtcVad},
     whisper::{process_with_whisper, WhisperModel},
     AudioDevice, AudioTranscriptionEngine,
 };
 use anyhow::{anyhow, Result};
 use candle_transformers::models::whisper as m;
-use hound::{WavSpec, WavWriter};
-use lazy_static::lazy_static;
 use log::{debug, error, info};
 #[cfg(target_os = "macos")]
 use objc::rc::autoreleasepool;
-use reqwest::Client;
 use screenpipe_core::Language;
-use serde_json::Value;
-use std::env;
-use std::io::Cursor;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::{
     path::PathBuf,
@@ -32,139 +22,8 @@ use std::{
     time::{SystemTime, UNIX_EPOCH},
 };
 use tokio::sync::Mutex;
-use vad_rs::VadStatus;
 
-lazy_static! {
-    static ref DEEPGRAM_API_URL: String = env::var("DEEPGRAM_API_URL")
-        .unwrap_or_else(|_| "https://api.deepgram.com/v1/listen".to_string());
-    static ref CUSTOM_DEEPGRAM_API_TOKEN: String =
-        env::var("CUSTOM_DEEPGRAM_API_TOKEN").unwrap_or_else(|_| String::new());
-}
-
-async fn transcribe_with_deepgram(
-    api_key: &str,
-    audio_data: &[f32],
-    device: &str,
-    sample_rate: u32,
-    languages: Vec<Language>,
-) -> Result<String> {
-    debug!("starting deepgram transcription");
-    let client = Client::new();
-
-    // Use token from env var
-    let custom_api_key = CUSTOM_DEEPGRAM_API_TOKEN.as_str();
-
-    // Create a WAV file in memory
-    let mut cursor = Cursor::new(Vec::new());
-    {
-        let spec = WavSpec {
-            channels: 1,
-            sample_rate: match sample_rate {
-                88200 => 16000,   // Deepgram expects 16kHz for 88.2kHz
-                _ => sample_rate, // Fallback for other sample rates
-            },
-            bits_per_sample: 32,
-            sample_format: hound::SampleFormat::Float,
-        };
-        let mut writer = WavWriter::new(&mut cursor, spec)?;
-        for &sample in audio_data {
-            writer.write_sample(sample)?;
-        }
-        writer.finalize()?;
-    }
-
-    // Get the WAV data from the cursor
-    let wav_data = cursor.into_inner();
-
-    let mut query_params = String::from("model=nova-2&smart_format=true&sample_rate=16000");
-
-    if !languages.is_empty() {
-        query_params = [
-            query_params,
-            "&".into(),
-            languages
-                .iter()
-                .map(|lang| format!("detect_language={}", lang.as_lang_code()))
-                .collect::<Vec<String>>()
-                .join("&"),
-        ]
-        .concat();
-    }
-
-    // rationale: custom api key = custom AI proxy to use deepgram
-    // no custom api key = use deepgram api key for real deepgram endpoint
-    let api_key_to_use = if custom_api_key.is_empty() {
-        api_key
-    } else {
-        custom_api_key
-    };
-    let is_custom_endpoint = !custom_api_key.is_empty();
-
-    debug!("deepgram api key: {}", api_key_to_use);
-
-    let response = client
-        .post(format!("{}?{}", *DEEPGRAM_API_URL, query_params))
-        .header("Content-Type", "audio/wav")
-        // Use Bearer format when using custom endpoint/proxy
-        .header(
-            "Authorization",
-            if is_custom_endpoint {
-                format!("Bearer {}", api_key_to_use)
-            } else {
-                format!("Token {}", api_key_to_use)
-            },
-        )
-        .body(wav_data)
-        .send();
-
-    match response.await {
-        Ok(resp) => {
-            debug!("received response from deepgram api");
-            match resp.json::<Value>().await {
-                Ok(result) => {
-                    debug!("successfully parsed json response");
-                    if let Some(err_code) = result.get("err_code") {
-                        error!(
-                            "deepgram api error code: {:?}, result: {:?}",
-                            err_code, result
-                        );
-                        return Err(anyhow::anyhow!("Deepgram API error: {:?}", result));
-                    }
-                    let transcription = result["results"]["channels"][0]["alternatives"][0]
-                        ["transcript"]
-                        .as_str()
-                        .unwrap_or("");
-
-                    if transcription.is_empty() {
-                        info!(
-                            "device: {}, transcription is empty. full response: {:?}",
-                            device, result
-                        );
-                    } else {
-                        info!(
-                            "device: {}, transcription successful. length: {} characters",
-                            device,
-                            transcription.len()
-                        );
-                    }
-
-                    Ok(transcription.to_string())
-                }
-                Err(e) => {
-                    error!("Failed to parse JSON response: {:?}", e);
-                    Err(anyhow::anyhow!("Failed to parse JSON response: {:?}", e))
-                }
-            }
-        }
-        Err(e) => {
-            error!("Failed to send request to Deepgram API: {:?}", e);
-            Err(anyhow::anyhow!(
-                "Failed to send request to Deepgram API: {:?}",
-                e
-            ))
-        }
-    }
-}
+pub use crate::segments::prepare_segments;
 
 pub fn stt_sync(
     audio: &[f32],
@@ -194,76 +53,6 @@ pub fn stt_sync(
     });
 
     handle.join().unwrap()
-}
-
-pub async fn prepare_segments(
-    audio_data: &[f32],
-    vad_engine: Arc<Mutex<Box<dyn VadEngine + Send>>>,
-    segmentation_model_path: &PathBuf,
-    embedding_manager: EmbeddingManager,
-    embedding_extractor: Arc<StdMutex<EmbeddingExtractor>>,
-    device: &str,
-) -> Result<tokio::sync::mpsc::Receiver<SpeechSegment>> {
-    let audio_data = normalize_v2(audio_data);
-
-    let frame_size = 1600;
-    let vad_engine = vad_engine.clone();
-
-    let mut noise = 0.;
-    let mut audio_frames = Vec::new();
-    let mut total_frames = 0;
-    let mut speech_frame_count = 0;
-
-    for chunk in audio_data.chunks(frame_size) {
-        total_frames += 1;
-
-        let mut new_chunk = chunk.to_vec();
-        let status = vad_engine.lock().await.audio_type(chunk);
-        match status {
-            Ok(VadStatus::Speech) => {
-                if let Ok(processed_audio) = spectral_subtraction(chunk, noise) {
-                    new_chunk = processed_audio;
-                    speech_frame_count += 1;
-                }
-            }
-            Ok(VadStatus::Unknown) => {
-                noise = average_noise_spectrum(chunk);
-            }
-            _ => {}
-        }
-        audio_frames.extend(new_chunk);
-    }
-
-    let speech_ratio = speech_frame_count as f32 / total_frames as f32;
-    let min_speech_ratio = vad_engine.lock().await.get_min_speech_ratio();
-
-    info!(
-        "device: {}, speech ratio: {}, min_speech_ratio: {}, audio_frames: {}, speech_frames: {}",
-        device,
-        speech_ratio,
-        min_speech_ratio,
-        audio_frames.len(),
-        speech_frame_count
-    );
-    let (tx, rx) = tokio::sync::mpsc::channel(100);
-    if !audio_frames.is_empty() && speech_ratio >= min_speech_ratio {
-        let segments = get_segments(
-            &audio_data,
-            16000,
-            segmentation_model_path,
-            embedding_extractor,
-            embedding_manager,
-        )?;
-
-        for segment in segments.flatten() {
-            if let Err(e) = tx.send(segment).await {
-                error!("failed to send segment: {:?}", e);
-                break;
-            }
-        }
-    }
-
-    Ok(rx)
 }
 
 #[allow(clippy::too_many_arguments)]
